add auth to chroma
