+++ packages/components/nodes/vectorstores/Chroma/Chroma_Existing.ts
@@ -1,8 +1,8 @@
 import { ICommonObject, INode, INodeData, INodeOutputsValue, INodeParams } from '../../../src/Interface'
-import { Chroma, ChromaLibArgs } from 'langchain/vectorstores/chroma'
+import { Chroma } from 'langchain/vectorstores/chroma'
 import { Embeddings } from 'langchain/embeddings/base'
 import { getBaseClasses, getCredentialData, getCredentialParam } from '../../../src/utils'
-import type { Collection } from 'chromadb'
+import { ChromaExtended } from './core'
 
 class Chroma_Existing_VectorStores implements INode {
     label: string
@@ -94,7 +94,7 @@ class Chroma_Existing_VectorStores implements INode {
         if (chromaURL) obj.url = chromaURL
         if (chromaApiKey) obj.chromaApiKey = chromaApiKey
 
-        const vectorStore = await ChromaExisting.fromExistingCollection(embeddings, obj)
+        const vectorStore = await ChromaExtended.fromExistingCollection(embeddings, obj)
 
         if (output === 'retriever') {
             const retriever = vectorStore.asRetriever(k)
@@ -104,52 +104,6 @@ class Chroma_Existing_VectorStores implements INode {
             return vectorStore
         }
         return vectorStore
-    }
-}
-
-interface ChromaAuth {
-    chromaApiKey?: string
-}
-
-class ChromaExisting extends Chroma {
-    chromaApiKey?: string
-
-    constructor(embeddings: Embeddings, args: ChromaLibArgs & Partial<ChromaAuth>) {
-        super(embeddings, args)
-        this.chromaApiKey = args.chromaApiKey
-    }
-
-    static async fromExistingCollection(embeddings: Embeddings, dbConfig: ChromaLibArgs & Partial<ChromaAuth>): Promise<Chroma> {
-        const instance = new this(embeddings, dbConfig)
-        await instance.ensureCollection()
-        return instance
-    }
-
-    async ensureCollection(): Promise<Collection> {
-        if (!this.collection) {
-            if (!this.index) {
-                const { ChromaClient } = await Chroma.imports()
-                const obj: any = {
-                    path: this.url
-                }
-                if (this.chromaApiKey) {
-                    obj.fetchOptions = {
-                        headers: {
-                            'X-Api-Key': this.chromaApiKey
-                        }
-                    }
-                }
-                this.index = new ChromaClient(obj)
-            }
-            try {
-                this.collection = await this.index.getOrCreateCollection({
-                    name: this.collectionName
-                })
-            } catch (err) {
-                throw new Error(`Chroma getOrCreateCollection error: ${err}`)
-            }
-        }
-        return this.collection
     }
 }
